Move layout to components folder and rename its import to be consistent
