+++ www/src/templates/template-blog-post.js
@@ -6,7 +6,7 @@ import ArrowBackIcon from "react-icons/lib/md/arrow-back"
 import Img from "gatsby-image"
 import { OutboundLink } from "gatsby-plugin-google-analytics"
 
-import GlobalLayout from "../layouts"
+import Layout from "../components/layout"
 import presets, { colors } from "../utils/presets"
 import typography, { rhythm, scale, options } from "../utils/typography"
 import Container from "../components/container"
@@ -60,7 +60,7 @@ class BlogPostTemplate extends React.Component {
     }
 
     return (
-      <GlobalLayout location={this.props.location}>
+      <Layout location={this.props.location}>
         <Container className="post" css={{ paddingBottom: `0 !important` }}>
           {/* Add long list of social meta tags */}
           <Helmet>
@@ -286,7 +286,7 @@ class BlogPostTemplate extends React.Component {
             </div>
           </Container>
         </div>
-      </GlobalLayout>
+      </Layout>
     )
   }
 }
